refactoring the service now formatters as per review comments
