+++ components/notifications-service/server/lib/data/target.ex
@@ -34,7 +34,7 @@ defmodule Notifications.TargetBuilder do
   defp formatter(Notifications.SlackAlert), do: &Formatters.Slack.format/1
   defp formatter(_), do: &Notifications.Target.null_format/1
 
-  defp criticals_only_formatter(Notifications.ServiceNowAlert), do: &Formatters.ServiceNowCritical.Compliance.format/1
+  defp criticals_only_formatter(Notifications.ServiceNowAlert), do: &Formatters.ServiceNow.Compliance.format_critical/1
 
 
   defp pre_filter(Notifications.ServiceNowAlert, event), do: event != 2
