+++ components/notifications-service/server/lib/formatters/compliance_helper.ex
@@ -103,31 +103,4 @@ defmodule Notifications.Formatters.ComplianceHelper do
   # Anything that's not a failure gets excluded
   defp prune_results([_| results], acc), do: prune_results(results, acc)
 
-  @spec get_servicenow_compliance_notification(ComplianceFailure.t):: map()
-  def get_servicenow_compliance_notification(notification) do
-    get_servicenow_compliance_notification(notification, @crit_control_threshold)
-  end
-  def get_servicenow_compliance_notification(%ComplianceFailure{test_totals: totals} = notification, control_threshold) do
-    Utils.to_map(notification)
-
-    notification = prune_and_augment(notification, control_threshold)
-    %{
-      automate_fqdn: Notifications.Config.automate_fqdn,
-      failure_snippet: "InSpec found a critical control failure on [#{notification.node_name}](#{notification.compliance_url})",
-      automate_failure_url: notification.compliance_url,
-      node_name: notification.node_name,
-      node_uuid: notification.node_id,
-      number_of_critical_tests: totals.critical,
-      total_number_of_tests: totals.failed + totals.passed + totals.skipped,
-      total_number_of_failed_tests: totals.failed,
-      number_of_failed_critical_tests: totals.critical_failed,
-      total_number_of_passed_tests: totals.passed,
-      total_number_of_skipped_tests: totals.skipped,
-      inspec_version: notification.inspec_version,
-      failed_critical_profiles: Utils.to_map(notification.failed_profiles),
-      timestamp_utc: Utils.format_date_string(notification.timestamp),
-      end_time_utc: Utils.format_date_string(notification.end_time),
-      type: "compliance_failure"
-    }
-  end
 end
