--- components/notifications-service/server/lib/formatters/servicenowcritical.compliance.ex
@@ -1,15 +1,0 @@
-defmodule Notifications.Formatters.ServiceNowCritical.Compliance do
-  @moduledoc """
-  Build a message map from the compliance struct to be supplied to the ServiceNow url
-  """
-  require Logger
-  @behaviour Notifications.Formatters.Behavior
-
-  alias Notifications.ComplianceFailure
-  alias Notifications.Formatters.ComplianceHelper
-
-  @spec format(ComplianceFailure.t):: map()
-  def format(notification) do
-    ComplianceHelper.get_servicenow_compliance_notification(notification)
-  end
-end
